docSense/requirements.txt
flask
endee
sentence-transformers
PyPDF2
python-docx
groq
python-dotenv
requests